# Conflicts:
#	src/Nethermind/Nethermind.Evm/VirtualMachine.cs

+++ src/Nethermind/Nethermind.Evm/Instructions/EvmInstructions.Math2Param.cs
@@ -23,7 +23,7 @@ public static partial class EvmInstructions
         /// <summary>
         /// The gas cost for executing this math operation.
         /// </summary>
-        virtual static long GasCost => GasCostOf.VeryLow;
+        virtual static ulong GasCost => GasCostOf.VeryLow;
         /// <summary>
         /// Executes the math operation on two 256-bit operands.
         /// </summary>
@@ -107,7 +107,7 @@ public static partial class EvmInstructions
     /// </summary>
     public struct OpMul : IOpMath2Param
     {
-        public static long GasCost => GasCostOf.Low;
+        public static ulong GasCost => GasCostOf.Low;
         public static void Operation(in UInt256 a, in UInt256 b, out UInt256 result)
             => UInt256.Multiply(in a, in b, out result);
     }
@@ -118,7 +118,7 @@ public static partial class EvmInstructions
     /// </summary>
     public struct OpDiv : IOpMath2Param
     {
-        public static long GasCost => GasCostOf.Low;
+        public static ulong GasCost => GasCostOf.Low;
         public static void Operation(in UInt256 a, in UInt256 b, out UInt256 result)
         {
             if (b.IsZero)
@@ -142,7 +142,7 @@ public static partial class EvmInstructions
     /// </summary>
     public struct OpSDiv : IOpMath2Param
     {
-        public static long GasCost => GasCostOf.Low;
+        public static ulong GasCost => GasCostOf.Low;
         public static void Operation(in UInt256 a, in UInt256 b, out UInt256 result)
         {
             if (b.IsZero)
@@ -173,7 +173,7 @@ public static partial class EvmInstructions
     /// </summary>
     public struct OpMod : IOpMath2Param
     {
-        public static long GasCost => GasCostOf.Low;
+        public static ulong GasCost => GasCostOf.Low;
         public static void Operation(in UInt256 a, in UInt256 b, out UInt256 result)
         {
             if (b.IsZeroOrOne)
@@ -195,7 +195,7 @@ public static partial class EvmInstructions
     /// </summary>
     public struct OpSMod : IOpMath2Param
     {
-        public static long GasCost => GasCostOf.Low;
+        public static ulong GasCost => GasCostOf.Low;
         public static void Operation(in UInt256 a, in UInt256 b, out UInt256 result)
         {
             if (b.IsZeroOrOne)
@@ -291,7 +291,7 @@ public static partial class EvmInstructions
             return stack.PushOne<TTracingInst>();
         }
 
-        int expSize = 32 - leadingZeros;
+        ulong expSize = (ulong)(32 - leadingZeros);
         // Deduct gas proportional to the number of 32-byte words needed to represent the exponent.
         TGasPolicy.Consume(ref gas, vm.Spec.GasCosts.ExpByteCost * expSize);
